# AquaSense/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AquaSense API",
    "version": "1.0.0",
    "description": "API для керування AquaSense системою"
  },
  "host": "localhost:3000",
  "schemes": ["http"],
...
  "paths": {
    "/users/register": {
      "post": {
        "tags": ["Authorization"],
        "summary": "Реєстрація користувача",
        "description": "Створення нового користувача",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Дані нового користувача",
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "email": { "type": "string" },
                "password": { "type": "string" }
              },
              "required": ["username", "email", "password"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Користувача успішно створено"
          },
          "400": {
            "description": "Користувач вже існує"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["Authorization"],
        "summary": "Авторизація користувача",
        "description": "Вхід користувача",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Дані для входу",
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "password": { "type": "string" }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Авторизація успішна"
          },
          "401": {
            "description": "Невірні дані для входу"
          },
          "404": {
            "description": "Користувача не знайдено"
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Отримання користувача за id",
        "description": "Повертає інформацію про користувача за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id користувача"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "username": { "type": "string" },
                "email": { "type": "string" },
                "created_at": { "type": "string", "format": "date-time" }
              }
            }
          },
          "404": {
            "description": "Користувача не знайдено"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Оновлення даних користувача",
        "description": "Оновлює дані користувача за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id користувача для оновлення"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Нові дані користувача",
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "email": { "type": "string" }
              },
              "required": ["username", "email"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Дані користувача успішно оновлено"
          },
          "404": {
            "description": "Користувача не знайдено"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Видалення користувача",
        "description": "Видаляє користувача за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id користувача для видалення"
          }
        ],
        "responses": {
          "200": {
            "description": "Користувача успішно видалено"
          },
          "404": {
            "description": "Користувача не знайдено"
          }
        }
      }
    },
...
    "/aquariums/user/{userId}": {
      "get": {
        "tags": ["Aquariums manage"],
        "summary": "Отримати всі акваріуми для користувача за його id",
        "description": "Повертає список акваріумів, що належать конкретному користувачу за його id",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "description": "id користувача"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "user_id": { "type": "integer" },
                  "name": { "type": "string" },
                  "specification": { "type": "string" },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          },
          "404": {
            "description": "Користувача не знайдено або у користувача немає акваріумів"
          }
        }
      }
    },
...
  }
}